# lib: window-based flow control