@junhochoi 2021-09-10
This PR mostly improves client download performance.

- Network: 1Gbit link, 20ms latency
- Server: running quiche-server (master + #987 branch)
- Client:
master branch: quiche-client downloading 1G file: 23.397s
```
[2021-09-10T04:26:49.594639000Z INFO  quiche_apps::common] 1/1 response(s) received in 23.397971791s, closing...
```
this branch: quiche-client download 1G file: 10.062s
```
[2021-09-10T04:30:40.581162000Z INFO  quiche_apps::common] 1/1 response(s) received in 10.062495518s, closing...
```

For comparison, h2load with quic patch on the same client:
```
$ h2load --npn-list h3-29 -n 1 https://buster:8443/1000m.dat
finished in 9.53s, 0.10 req/s, 104.88MB/s
```